chore(wmg): refactor sort subcomponent
